requirements.txt
requests
fastapi
uvicorn
python_dotenv
psycopg2-binary